create classes for tests
